TODO/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
djangorestframework==3.9.2
djangorestframework-simplejwt==4.1.0
httpie==1.0.2
idna==2.8
Pygments==2.3.1
PyJWT==1.7.1
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
